requirements.txt
Django
djangorestframework
gunicorn
psycopg2-binary
httpx
requests
celery[sqlalchemy]
SQLAlchemy
kombu[sqlalchemy]